+++ src/main/kotlin/id/walt/vclib/VcUtils.kt
@@ -21,7 +21,7 @@ object VcUtils {
             else -> SignedJWT.parse(vc.jwt).jwtClaimsSet.issuer
         }
         is GaiaxCredential -> vc.issuer
-        is GaiaxSD -> vc.issuer
+        is GaiaxSelfDescription -> vc.issuer
         else -> {
             log.warn { "No getIssuer for ${vc.type.last()}!" }
             ""
@@ -37,7 +37,7 @@ object VcUtils {
         is VerifiableAttestation -> vc.credentialSubject!!.id
         is VerifiableAuthorization -> vc.credentialSubject.id
         is GaiaxCredential -> vc.credentialSubject.id
-        is GaiaxSD -> vc.credentialSubject.id
+        is GaiaxSelfDescription -> vc.credentialSubject.id
         is VerifiablePresentation -> vc.holder!!
         else -> {
             log.warn { "No getHolder for ${vc.type.last()}!" }
@@ -53,7 +53,7 @@ object VcUtils {
         is VerifiableAttestation -> vc.issuanceDate
         is VerifiableAuthorization -> vc.issuanceDate
         is GaiaxCredential -> vc.issuanceDate
-        is GaiaxSD -> vc.issuanceDate
+        is GaiaxSelfDescription -> vc.issuanceDate
         else -> {
             log.warn { "No getIssuanceDate for ${vc.type.last()}!" }
             ""
@@ -67,7 +67,7 @@ object VcUtils {
         is VerifiableAttestation -> vc.validFrom
         is VerifiableAuthorization -> vc.validFrom
         is GaiaxCredential -> vc.validFrom
-        is GaiaxSD -> vc.validFrom
+        is GaiaxSelfDescription -> vc.validFrom
         else -> {
             log.warn { "No getValidFrom for ${vc.type.last()}!" }
             ""
@@ -80,7 +80,7 @@ object VcUtils {
         is VerifiableDiploma -> vc.expirationDate
         is VerifiableAuthorization -> vc.expirationDate
         is GaiaxCredential -> vc.expirationDate
-        is GaiaxSD -> vc.expirationDate
+        is GaiaxSelfDescription -> vc.expirationDate
         else -> {
             log.warn { "No getExpirationDate for ${vc.type.last()}!" }
             ""
@@ -110,7 +110,7 @@ object VcUtils {
             is VerifiableAuthorization -> vc.proof!!.nonce
             is VerifiablePresentation -> vc.proof!!.nonce
             is GaiaxCredential -> vc.proof!!.nonce
-            is GaiaxSD -> vc.proof!!.nonce
+            is GaiaxSelfDescription -> vc.proof!!.nonce
             else -> {
                 log.warn { "No getCredentialSchema for ${vc.type.last()}!" }
                 null